WIP: philo - mandatory (nticks)

+++ circle_03/philosophers/philo/src/print.c
@@ -6,7 +6,7 @@
 /*   By: chaejkim <chaejkim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/08 13:54:57 by chaejkim          #+#    #+#             */
-/*   Updated: 2022/07/15 13:18:16 by chaejkim         ###   ########.fr       */
+/*   Updated: 2022/07/22 16:15:17 by chaejkim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 int	timestamp(t_simulation_info *sinfo, int philo_num, int state)
 {
 	static int	need_end = FALSE;
-	float		mticks;
+	int			mticks;
 	t_tv		current;
 
 	pthread_mutex_lock(&sinfo->printer);
@@ -25,17 +25,17 @@ int	timestamp(t_simulation_info *sinfo, int philo_num, int state)
 	gettimeofday(&current, NULL);
 	mticks = time_diff(&sinfo->start, &current);
 	if (state == TAKE)
-		printf("%f %d has taken a fork\n", mticks, philo_num);
+		printf("%d %d has taken a fork\n", mticks, philo_num);
 	else if (state == EAT)
-		printf("\033[0;33m%f %d is eating\n\033[0m", mticks, philo_num);
+		printf("\033[0;33m%d %d is eating\n\033[0m", mticks, philo_num);
 	else if (state == SLEEP)
-		printf("%f %d is sleeping\n", mticks, philo_num);
+		printf("%d %d is sleeping\n", mticks, philo_num);
 	else if (state == THINK)
-		printf("\033[0;32m%f %d is thinking\n\033[0m", mticks, philo_num);
+		printf("\033[0;32m%d %d is thinking\n\033[0m", mticks, philo_num);
 	else if (state == DEAD)
 	{
 		need_end = TRUE;
-		printf("\033[0;31m%f %d died\n\033[0m", mticks, philo_num);
+		printf("\033[0;31m%d %d died\n\033[0m", mticks, philo_num);
 	}
 	return (pthread_mutex_unlock(&sinfo->printer));
 }
